feat: refactor ClusterTopology management and add unit tests
